Pulling up common methods and fields.
