+++ core/src/main/java/uk/ac/shef/dcs/sti/experiment/BaselineBatch.java
@@ -109,7 +109,7 @@ public class BaselineBatch extends STIBatch {
                     , e);
         }
 
-        //LOG.info("Initializing WebSearcher...");
+        //log.info("Initializing WebSearcher...");
 
 
         LOG.info("Initializing SUBJECT COLUMN DETECTION components ...");
