+++ kbsearch/src/main/java/uk/ac/shef/dcs/kbsearch/sparql/SPARQLSearch.java
@@ -1,13 +1,21 @@
 package uk.ac.shef.dcs.kbsearch.sparql;
 
 import javafx.util.Pair;
+
+import org.apache.commons.lang3.StringEscapeUtils;
 import org.apache.jena.query.*;
 import org.apache.jena.rdf.model.RDFNode;
+import org.apache.log4j.Logger;
 import org.apache.solr.client.solrj.embedded.EmbeddedSolrServer;
 import org.simmetrics.StringMetric;
 import org.simmetrics.metrics.Levenshtein;
 import uk.ac.shef.dcs.kbsearch.KBSearch;
 import uk.ac.shef.dcs.kbsearch.KBSearchException;
+import uk.ac.shef.dcs.kbsearch.model.Attribute;
+import uk.ac.shef.dcs.kbsearch.model.Clazz;
+import uk.ac.shef.dcs.kbsearch.model.Entity;
+import uk.ac.shef.dcs.util.SolrCache;
+import uk.ac.shef.dcs.util.StringUtils;
 
 import java.io.IOException;
 import java.util.*;
@@ -33,7 +41,7 @@ import java.util.*;
  */
 public abstract class SPARQLSearch extends KBSearch {
 
-    protected String sparqlEndpoint;
+  protected String sparqlEndpoint;
 
     protected StringMetric stringMetric = new Levenshtein();
     /**
@@ -148,5 +156,309 @@ public abstract class SPARQLSearch extends KBSearch {
             Double s2 = scores.get(o2);
             return s2.compareTo(s1);
         });
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    public List<Entity> findEntityCandidates(String content) throws KBSearchException {
+        String query = createSolrCacheQuery_findResources(content);
+    
+        content = StringEscapeUtils.unescapeXml(content);
+        int bracket = content.indexOf("(");
+        if (bracket != -1) {
+            content = content.substring(0, bracket).trim();
+        }
+        if (StringUtils.toAlphaNumericWhitechar(content).trim().length() == 0)
+            return new ArrayList<>();
+    
+    
+        List<Entity> result = null;
+        if (!ALWAYS_CALL_REMOTE_SEARCHAPI) {
+            try {
+                result = (List<Entity>) cacheEntity.retrieve(query);
+                if (result != null)
+                    log.debug("QUERY (entities, cache load)=" + query + "|" + query);
+            } catch (Exception e) {
+                log.error(e.getLocalizedMessage(),e);
+            }
+        }
+        if (result == null) {
+            result = new ArrayList<>();
+            try {
+                //1. try exact string
+                String sparqlQuery = createExactMatchQueries(content);
+                List<Pair<String, String>> queryResult = queryByLabel(sparqlQuery, content);
+    
+                //2. if result is empty, try regex
+                if (queryResult.size() == 0 && fuzzyKeywords) {
+                    log.debug("(query by regex. This can take a long time)");
+                    sparqlQuery = createRegexQuery(content);
+                    queryResult = queryByLabel(sparqlQuery, content);
+                }
+                //3. rank result by the degree of matches
+                rank(queryResult, content);
+    
+                //get attributes for each resource
+                log.debug("(QUERY =" + queryResult.size() + " results)");
+                for (Pair<String, String> candidate : queryResult) {
+                    String label = candidate.getValue();
+                    if (label == null)
+                        label = content;
+                    Entity ec = new Entity(candidate.getKey(), label);
+                    List<Attribute> attributes = findAttributesOfEntities(ec);
+                    ec.setAttributes(attributes);
+                    for (Attribute attr : attributes) {
+                        adjustValueOfURLResource(attr);
+                        if (attr.getRelationURI().endsWith(RDFEnum.RELATION_HASTYPE_SUFFIX_PATTERN.getString()) &&
+                                !ec.hasType(attr.getValueURI())) {
+                            ec.addType(new Clazz(attr.getValueURI(), attr.getValue()));
+                        }
+                    }
+                    result.add(ec);
+                }
+    
+                cacheEntity.cache(query, result, AUTO_COMMIT);
+                log.debug("QUERY (entities, cache save)=" + query + "|" + query);
+            } catch (Exception e) {
+                throw new KBSearchException(e);
+            }
+        }
+    
+        //filter entity's clazz, and attributes
+        String id = "|";
+        for (Entity ec : result) {
+            id = id + ec.getId() + ",";
+            //ec.setTypes(FreebaseSearchResultFilter.filterClazz(ec.getTypes()));
+            List<Clazz> filteredTypes = resultFilter.filterClazz(ec.getTypes());
+            ec.clearTypes();
+            for (Clazz ft : filteredTypes)
+                ec.addType(ft);
+        }
+    
+        return result;
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
+    public List<Entity> findEntityCandidatesOfTypes(String content, String... types) throws KBSearchException {
+        String queryCache = createSolrCacheQuery_findResources(content);
+    
+        content = StringEscapeUtils.unescapeXml(content);
+        int bracket = content.indexOf("(");
+        if (bracket != -1) {
+            content = content.substring(0, bracket).trim();
+        }
+        if (StringUtils.toAlphaNumericWhitechar(content).trim().length() == 0)
+            return new ArrayList<>();
+    
+    
+        List<Entity> result = null;
+        if (!ALWAYS_CALL_REMOTE_SEARCHAPI) {
+            try {
+                result = (List<Entity>) cacheEntity.retrieve(queryCache);
+                if (result != null) {
+                    log.debug("QUERY (entities, cache load)=" + queryCache + "|" + queryCache);
+                    if (types.length > 0) {
+                        Iterator<Entity> it = result.iterator();
+                        while (it.hasNext()) {
+                            Entity ec = it.next();
+                            boolean typeSatisfied = false;
+                            for (String t : types) {
+                                if (ec.hasType(t)) {
+                                    typeSatisfied = true;
+                                    break;
+                                }
+                            }
+                            if (!typeSatisfied)
+                                it.remove();
+                        }
+                    }
+                }
+            } catch (Exception e) {
+                log.error(e.getLocalizedMessage(),e);
+            }
+        }
+        if (result == null) {
+            result = new ArrayList<>();
+            try {
+                //1. try exact string
+                String sparqlQuery = createExactMatchWithOptionalTypes(content);
+                List<Pair<String, String>> resourceAndType = queryByLabel(sparqlQuery, content);
+                boolean hasExactMatch = resourceAndType.size() > 0;
+                if (types.length > 0) {
+                    Iterator<Pair<String, String>> it = resourceAndType.iterator();
+                    while (it.hasNext()) {
+                        Pair<String, String> ec = it.next();
+                        boolean typeSatisfied = false;
+                        for (String t : types) {
+                            if (t.equals(ec.getValue())) {
+                                typeSatisfied = true;
+                                break;
+                            }
+                        }
+                        if (!typeSatisfied)
+                            it.remove();
+                    }
+                }//with this query the 'value' of the pair will be the type, now need to reset it to actual value
+                List<Pair<String, String>> queryResult = new ArrayList<>();
+                if (resourceAndType.size() > 0) {
+                    Pair<String, String> matchedResource = resourceAndType.get(0);
+                    queryResult.add(new Pair<>(matchedResource.getKey(), content));
+                }
+    
+                //2. if result is empty, try regex
+                if (!hasExactMatch && fuzzyKeywords) {
+                    log.debug("(query by regex. This can take a long time)");
+                    sparqlQuery = createRegexQuery(content, types);
+                    queryResult = queryByLabel(sparqlQuery, content);
+                }
+                //3. rank result by the degree of matches
+                rank(queryResult, content);
+    
+                //firstly fetch candidate freebase topics. pass 'true' to only keep candidates whose name overlap with the query term
+                log.debug("(DB QUERY =" + queryResult.size() + " results)");
+                for (Pair<String, String> candidate : queryResult) {
+                    //Next get attributes for each topic
+                    String label = candidate.getValue();
+                    if (label == null)
+                        label = content;
+                    Entity ec = new Entity(candidate.getKey(), label);
+                    List<Attribute> attributes = findAttributesOfEntities(ec);
+                    ec.setAttributes(attributes);
+                    for (Attribute attr : attributes) {
+                        adjustValueOfURLResource(attr);
+                        if (attr.getRelationURI().endsWith(RDFEnum.RELATION_HASTYPE_SUFFIX_PATTERN.getString()) &&
+                                !ec.hasType(attr.getValueURI())) {
+                            ec.addType(new Clazz(attr.getValueURI(), attr.getValue()));
+                        }
+                    }
+                    result.add(ec);
+                }
+    
+                cacheEntity.cache(queryCache, result, AUTO_COMMIT);
+                log.debug("QUERY (entities, cache save)=" + queryCache + "|" + queryCache);
+            } catch (Exception e) {
+                throw new KBSearchException(e);
+            }
+        }
+    
+        //filter entity's clazz, and attributes
+        String id = "|";
+        for (Entity ec : result) {
+            id = id + ec.getId() + ",";
+            //ec.setTypes(FreebaseSearchResultFilter.filterClazz(ec.getTypes()));
+            List<Clazz> filteredTypes = resultFilter.filterClazz(ec.getTypes());
+            ec.clearTypes();
+            for (Clazz ft : filteredTypes)
+                ec.addType(ft);
+        }
+    
+        return result;
+    }
+
+    private void adjustValueOfURLResource(Attribute attr) throws KBSearchException {
+        String value = attr.getValue();
+        if (value.startsWith("http")) {
+            String queryCache = createSolrCacheQuery_findLabelForResource(value);
+    
+    
+            List<String> result = null;
+            if (!ALWAYS_CALL_REMOTE_SEARCHAPI) {
+                try {
+                    result = (List<String>) cacheEntity.retrieve(queryCache);
+                    if (result != null) {
+                        log.debug("QUERY (resource labels, cache load)=" + queryCache + "|" + queryCache);
+                    }
+                } catch (Exception e) {
+                }
+            }
+            if (result == null) {
+                try {
+                    //1. try exact string
+                    String sparqlQuery = createGetLabelQuery(value);
+                    result = queryForLabel(sparqlQuery, value);
+    
+                    cacheEntity.cache(queryCache, result, AUTO_COMMIT);
+                    log.debug("QUERY (entities, cache save)=" + queryCache + "|" + queryCache);
+                } catch (Exception e) {
+                    throw new KBSearchException(e);
+                }
+            }
+    
+            if (result.size() > 0) {
+                attr.setValueURI(value);
+                attr.setValue(result.get(0));
+            } else {
+                attr.setValueURI(value);
+            }
+        }
+    }
+
+    @Override
+    public List<Attribute> findAttributesOfEntities(Entity ec) throws KBSearchException {
+        return findAttributes(ec.getId(), cacheEntity);
+    }
+
+    private List<Attribute> findAttributes(String id, SolrCache cache) throws KBSearchException {
+        if (id.length() == 0)
+            return new ArrayList<>();
+    
+        String queryCache = createSolrCacheQuery_findAttributesOfResource(id);
+        List<Attribute> result = null;
+        if (!ALWAYS_CALL_REMOTE_SEARCHAPI) {
+            try {
+                result = (List<Attribute>) cache.retrieve(queryCache);
+                if (result != null)
+                    log.debug("QUERY (attributes of id, cache load)=" + queryCache + "|" + queryCache);
+            } catch (Exception e) {
+                log.error(e.getLocalizedMessage(), e);
+            }
+        }
+    
+        if (result == null) {
+            result = new ArrayList<>();
+            String query = "SELECT DISTINCT ?p ?o WHERE {\n" +
+                    "<" + id + "> ?p ?o .\n" +
+                    "}";
+    
+            Query sparqlQuery = QueryFactory.create(query);
+            QueryExecution qexec = QueryExecutionFactory.sparqlService(sparqlEndpoint, sparqlQuery);
+    
+            ResultSet rs = qexec.execSelect();
+            while (rs.hasNext()) {
+                QuerySolution qs = rs.next();
+                RDFNode predicate = qs.get("?p");
+                RDFNode object = qs.get("?o");
+                if (object != null) {
+                    Attribute attr = new DBpediaAttribute(predicate.toString(), object.toString());
+                    result.add(attr);
+                }
+            }
+    
+            try {
+                cache.cache(queryCache, result, AUTO_COMMIT);
+                log.debug("QUERY (attributes of id, cache save)=" + query + "|" + query);
+            } catch (Exception e) {
+                log.error(e.getLocalizedMessage(), e);
+            }
+        }
+    
+        //filtering
+        result = resultFilter.filterAttribute(result);
+        return result;
+    }
+
+    @Override
+    public List<Attribute> findAttributesOfClazz(String clazzId) throws KBSearchException {
+        return findAttributes(clazzId, cacheEntity);
+    }
+
+    @Override
+    public List<Attribute> findAttributesOfProperty(String propertyId) throws KBSearchException {
+        return findAttributes(propertyId, cacheEntity);
+    }
+
+    protected String createSolrCacheQuery_findLabelForResource(String url) {
+        return "LABEL_" + url;
     }
 }
